+++ crates/pipeline-manager/src/api/main.rs
@@ -438,6 +438,7 @@ It contains the following fields:
         // Adapter statistics
         feldera_types::adapter_stats::ExternalControllerStatus,
         feldera_types::adapter_stats::ExternalGlobalControllerMetrics,
+        feldera_types::adapter_stats::ConnectorError,
         feldera_types::adapter_stats::ExternalInputEndpointStatus,
         feldera_types::adapter_stats::ExternalInputEndpointMetrics,
         feldera_types::adapter_stats::ExternalOutputEndpointStatus,